backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.1.1
qdrant-client==1.11.3
PyMuPDF>=1.25.1
groq>=0.23.0
google-generativeai==0.8.3
pydantic==2.9.2
requests
beautifulsoup4
apscheduler==3.10.4
python-dotenv
python-multipart
transformers>=4.36.0
torch>=2.1.0
